fix: version info in artifact links and emoji and security patches
