+++ public/package.json
@@ -23,7 +23,7 @@
     "@sentry/tracing": "^6.2.3",
     "bootstrap": "^4.4.1",
     "compare-versions": "^3.6.0",
-    "emoji-regex": "^9.2.2",
+    "emoji-regex": "^10.0.0",
     "jquery": "^3.5.0",
     "popper.js": "^1.16.0",
     "rxjs": "~6.6.7",
